backend/requirements.txt
Django==5.1.4
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.3.1
requests==2.32.5
twilio==9.4.0
django-allauth==65.3.0
dj-rest-auth==7.0.0
cryptography
Pillow==10.4.0
